FF-328 #comment placeholder. Added HistoryAndTriggers to app.js
